Allow to skip base LDIF import
In order to be a bit more flexible when testing against libregraph-idm
introduce the new variable REVA_LDAP_SKIP_LDIF_IMPORT to allow to skip
the import of the base LDIF file that creates the ou objects for users
and groups (these objects are already present in such a setup).

This also means that when removing users and groups after each test we
cannot rely on recursivly removing the ou object about explicitly delete
every user/group.

+++ tests/acceptance/features/bootstrap/Provisioning.php
@@ -539,6 +539,7 @@ trait Provisioning {
 			$useSsl = OcisHelper::useSsl();
 			$this->ldapAdminUser = OcisHelper::getBindDN();
 			$this->ldapAdminPassword = OcisHelper::getBindPassword();
+			$this->skipImportLdif = (\getenv("REVA_LDAP_SKIP_LDIF_IMPORT") === "true");
 			if ($useSsl === true) {
 				\putenv('LDAPTLS_REQCERT=never');
 			}
@@ -596,7 +597,9 @@ trait Provisioning {
 				$ldifFile = $configPath . "/" . \basename($ldifFile);
 			}
 		}
-		$this->importLdifFile($ldifFile);
+		if (!$this->skipImportLdif) {
+			$this->importLdifFile($ldifFile);
+		}
 		$this->theLdapUsersHaveBeenResynced();
 	}
 
@@ -835,21 +838,29 @@ trait Provisioning {
 	 * @throws Exception
 	 */
 	public function deleteLdapUsersAndGroups():void {
-		//delete created ldap users
-		$this->ldap->delete(
-			"ou=" . $this->ldapUsersOU . "," . $this->ldapBaseDN,
-			true
-		);
-		//delete all created ldap groups
-		$this->ldap->delete(
-			"ou=" . $this->ldapGroupsOU . "," . $this->ldapBaseDN,
-			true
-		);
 		foreach ($this->ldapCreatedUsers as $user) {
+			$this->ldap->delete(
+				"uid=" . ldap_escape($user, "", LDAP_ESCAPE_DN) . ",ou=" . $this->ldapUsersOU . "," . $this->ldapBaseDN,
+			);
 			$this->rememberThatUserIsNotExpectedToExist($user);
 		}
 		foreach ($this->ldapCreatedGroups as $group) {
+			$this->ldap->delete(
+				"cn=" . ldap_escape($group, "", LDAP_ESCAPE_DN) . ",ou=" . $this->ldapGroupsOU . "," . $this->ldapBaseDN,
+			);
 			$this->rememberThatGroupIsNotExpectedToExist($group);
+		}
+		if (!$this->skipImportLdif) {
+			//delete ou from LDIF import
+			$this->ldap->delete(
+				"ou=" . $this->ldapUsersOU . "," . $this->ldapBaseDN,
+				true
+			);
+			//delete all created ldap groups
+			$this->ldap->delete(
+				"ou=" . $this->ldapGroupsOU . "," . $this->ldapBaseDN,
+				true
+			);
 		}
 		$this->theLdapUsersHaveBeenResynced();
 	}
